requirement.txt
abupy==0.4.0
alembic==0.9.6
aniso8601==2.0.0
antiorm==1.2.1
appdirs==1.4.3
apptools==4.4.0
arrow==0.12.0
asn1crypto==0.23.0
attrs==17.2.0
Automat==0.6.0
baidu-aip==1.6.4.0
bcolz==0.12.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
bitarray==0.8.1
bleach==1.5.0
blinker==1.4
Bottleneck==1.2.1
bs4==0.0.1
captcha==0.2.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
codegen==1.0
colorama==0.3.9
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.1.1
cssselect==1.0.1
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.27.3
db==0.1.1
db-sqlite3==0.0.1
decorator==4.1.2
Django==1.11.5
dnspython==1.3.4
dominate==2.3.1
empyrical==0.3.3
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
Faker==0.8.8
Flask==0.10.1
flask-blueprint==1.2.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.1
Flask-Dropzone==1.4.1
Flask-Email==1.4.4
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
gevent==1.2.2
graphviz==0.8.1
greenlet==0.4.12
gunicorn==19.7.1
h5py==2.7.1
hmmlearn==0.2.1
html5lib==0.9999999
hyperlink==17.3.1
idna==2.6
image==1.5.15
incremental==17.5.0
intervaltree==2.1.0
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
itchat==1.3.9
itsdangerous==0.24
jdcal==1.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyter-pip==0.3.1
Keras==2.1.1
llvmlite==0.21.0
Logbook==1.1.0
lru-dict==1.1.6
lxml==3.7.3
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mayavi==4.5.0+vtk71
mistune==0.7.4
multipledispatch==0.4.9
mxnet==1.0.0.post1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.2.0
numba==0.36.1
numexpr==2.6.4
numpy==1.13.3+mkl
olefile==0.44
opencv-python==3.3.1
openpyxl==2.4.8
optionaldict==0.1.1
ordereddict==1.1
pandas==0.18.1
pandas-datareader==0.5.0
pandocfilters==1.4.2
paramiko==2.3.1
parse==1.8.2
parsel==1.2.0
parso==0.1.0
patsy==0.4.1
patterns==0.3
pexpect==4.3.0
php==1.2.0
pickleshare==0.7.4
Pillow==4.2.1
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
py==1.5.2
PyAlgoTrade==0.18
pyasn1==0.3.7
pyasn1-modules==0.1.5
pybloom==2.0.0
pycparser==2.18
pycuda==2017.1.1+cuda90176
pycurl==7.43.0
PyDispatcher==2.0.5
pyecharts==0.2.7
pyface==5.1.0
pygame==1.9.3
Pygments==2.2.0
pyHook==1.5.1
pymc==2.3.6
pymongo==3.4.0
PyMySQL==0.7.11
PyNaCl==1.2.0
pyopencl==2017.2+cl21
PyOpenGL==3.1.1
PyOpenGL-accelerate==3.1.1
pyOpenSSL==17.3.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
PyQt5==5.9.1
pytest==3.3.1
python-dateutil==2.6.1
python-docx==0.8.6
python-editor==1.0.3
pytools==2017.6
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
QuantLib-Python==1.10
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
scikit-fmm==0.0.9
scikit-image==0.13.1
scikit-learn==0.19.1
scikits.odes==2.3.1
scikits.vectorplot==0.1.1
scipy==1.0.0
Scrapy==1.4.0
scrapy-redis==0.6.8
scrapyd==1.2.0
scrapyd-client==1.1.0
seaborn==0.8.1
selenium==3.5.0
service-identity==17.0.0
simplegeneric==0.8.1
SimpleITK==1.0.0
sip==4.19.5
six==1.10.0
sortedcontainers==1.5.9
spliter==1.0.0
SQLAlchemy==1.1.15
statsmodels==0.8.0
TA-Lib==0.4.10
tables==3.4.2
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
testpath==0.3.1
text-unidecode==1.1
Theano==0.9.0
threadpool==1.3.2
tifffile==2017.9.29
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
traits==4.6.0
traitsui==5.1.0
tushare==1.0.5
Twisted==17.9.0
urllib3==1.22
VideoCapture==0.9.5
visitor==0.1.3
VTK==8.0.1
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
wechatpy==1.6.0
Werkzeug==0.12.2
widgetsnbextension==3.0.6
win32compat==221.26
win32core==221.35
wordcloud==1.3.2
WTForms==2.1
xgboost==0.6+20171121
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
xmltodict==0.11.0
you-get==0.4.939
zipline==1.1.1
zope.interface==4.4.3
